backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
python-multipart==0.0.6
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

